Exp opt  merge to master; check primelib cpp for correct brackets

+++ PrimeFactorizer/PrimeCheck.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "pch.h"
+#include "STDlibs.h"
 
 using namespace std;
 
@@ -9,7 +9,7 @@ class PrimeCheck
 public:
 	PrimeCheck() noexcept; 
 	PrimeCheck(shared_ptr <PrimeLibrary> l);
-	bool is_prime(unsigned long long input);
+	bool is_prime(uint64_t input);
 
 private:
 	shared_ptr <PrimeLibrary> Lib;
