+++ PrimeFactorizer/PrimeFactor.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "pch.h"
+#include "STDlibs.h"
 #include "PrimeCheck.h"
 #include "PrimeDivisor.h"
 
@@ -11,13 +11,13 @@ public:
 	PrimeFactor() ;//no lib
 	PrimeFactor(shared_ptr <PrimeLibrary> l);//0-4; select library size with 0 being no library and 4 being the largest
 	~PrimeFactor();
-	vector <unsigned long long> factor(unsigned long long num);
+	vector <uint64_t> factor(uint64_t num);
 
 private:
 
 	unique_ptr <PrimeCheck> Check;
 	unique_ptr <PrimeDivisor> Div;
-	vector <unsigned long long> Divisors;
+	vector <uint64_t> Divisors;
 
 };
